fix: disable sentry on worker (#159)
* fix: disable sentry on worker

* fix: json

* fix: comment
