+++ examples/.env.sample
@@ -13,4 +13,4 @@ TG_BISCUICUITS_GITHUB_CLIENT_SECRET=fake
 TG_ELASTIC_CERT="-----BEGIN CERTIFICATE-----\nMIIFWTCCA0GgAwIBAgIUZWnGpkvJcphtt0jQKN9yIPw9CKEwDQYJKoZIhvcNAQEL\nBQAwPDE6MDgGA1UEAxMxRWxhc3RpY3NlYXJjaCBzZWN1cml0eSBhdXRvLWNvbmZp\nZ3VyYXRpb24gSFRUUCBDQTAeFw0yMjEwMjUxNDUxMDdaFw0yNTEwMjQxNDUxMDda\nMDwxOjA4BgNVBAMTMUVsYXN0aWNzZWFyY2ggc2VjdXJpdHkgYXV0by1jb25maWd1\ncmF0aW9uIEhUVFAgQ0EwggIiMA0GCSqGSIb3DQEBAQUAA4ICDwAwggIKAoICAQDE\nbVsMVOaPntg9R/8gB8idAXNKr4I+lOKyIk4c9eW/RoRpaoaZyFh68pG1IbV3MZLx\nWSSHKf/Q9Dy84OufBJ9PwqWpDix7F+L4ABgfw0GPPKm1xAuAVkFT9jtbT0NsPQsm\nv0hBH83hvLBCFOr0q6fG2/vQbRMkkw2fZKo3k7TOn89YDNKGh0g3/Yu4S7IAEyJw\nQsCJGxjixE700QWAoEcJ/7Ey5h9DQJZ9xQs4JoqfsDC1w40w6Fann3kR2qsWGyJ0\nQKwOmeagv7zc0qADdXfeuAJAy2PSxNdZUUgpomlYL48LjgQxuCp3aRRA5PKnu0pY\nw7GzKyc9Td/FD3jb1mIqb1bNqMmweWH9tPxFeSfNYuGTwQQWtcl7qBqWrDr08MB/\nKFkdbtwjewdTzBVNrSlbxKL7raEMJFubUZYYqM32tE3EfY5jm7TH3TWO+WD1yejh\nOxVf6S9ZarFJtdXi2weJHXXo85yxW+IAQxEKf1ojEx7Ju5hrfh+XoLWxfot3kHk2\n/BNyrlw5QekO6+IVrR7d3SM1slU8/UB/FC5HZYeZl+Wc0/4ICWYPRK7HPy4+DZUU\nHPXbWSZTRSJT/Z9TYFLofbtVbKZHIHAvDN9a0iTQINSHr2DroaG/KE6NtRXLfydZ\nl4A6ElzmWCkl3YfHzzchZKPVDpDd+qF7U0euoJNJPQIDAQABo1MwUTAdBgNVHQ4E\nFgQUwABfnsldWZacqMWth3C8Y42iQhswHwYDVR0jBBgwFoAUwABfnsldWZacqMWt\nh3C8Y42iQhswDwYDVR0TAQH/BAUwAwEB/zANBgkqhkiG9w0BAQsFAAOCAgEAciEv\nmM3HUMdVSjWwZMAqr07UgLGGbHFfPJQweiCMNaah+HFRrOEyEru6NXeoFieHcgbn\nB1H332eJTy47uBJKAvbYwCHMe68Ll2an2OB6Vn30WrickrIWh6mPH3geeIwOU6g+\nxUdfmyiiWdPBZzne4PWr9qWNmoZYnVcC4o05NZHYH6mBXRFq4wZCVRL0TkO/eozO\nL3OnekGkevo5fhH3oUvUIb8ZGcPi8Uny8gcl2lNfy/WFlfieybF03adFPf0zC1J5\no/I65bT/NKumMtYdJvyHL+WfRpnkwjPDFCLb8hNWI0CLqDpAJaLX6AfLDlJJYuT5\nXXroMiFxioVHVoBkyJDSawRzvJa62bhhznBUSR9gYgN/XnDkbeDshQCYS4UZZFXv\nBwgDKKI+m2PkdzcqZqbpIVV+ezubsJ+gQIg5K9fRhyw7BCN/vPU23TjxQ/5Ns/Ul\npCR/bHAW9FEy/hM7qXVedmv9tJ2zOFXd/5FUn8wvQFakBBBLKqnY+ueLSCYf1g/F\n4Ex0VrZRy7pBY2/w2RSng6LEnBwRLjJnFxrvVsk9Ciqju27F7FGiHcNel8M8GiMW\n88+UN5hVddCHwVsFiiB/dGcHiAtsN3FQEkxkMWvo4PCuIU+RXPhjfr11xw0Q2F7t\nR5U+Of7xVw43UYt+2ogd87hZNMM82w6EWYICJdY=\n-----END CERTIFICATE-----"
 TG_ELASTIC_BASIC="elastic:password"
 
-TG_BLOG_POSTGRES="postgresql://postgres:password@localhost:5432/db?schema=blog"
+TG_BLOG_POSTGRES="postgresql://postgres:password@host.docker.internal:5432/db?schema=blog"
