+++ examples/docker-compose.yml
@@ -1,12 +1,17 @@
 services:
   typegate:
     image: ghcr.io/metatypedev/typegate:v0.0.3-dev.5
+    platform: linux/amd64
+    #build:
+    #  context: ..
+    #  dockerfile: dev/Dockerfile
     restart: always
     ports:
       - "7890:7890"
     extra_hosts:
       - "host.docker.internal:host-gateway"
     environment:
+      DEBUG: "true"
       REDIS_URL: redis://:password@redis:6379/0
       TG_EXTERNAL_URL: http://localhost:7890
       TG_SECRET: "a4lNi0PbEItlFZbus1oeH/+wyIxi9uH6TpL8AIqIaMBNvp7SESmuUBbfUwC0prxhGhZqHw8vMDYZAGMhSZ4fLw=="
