+++ typegate/deno.json
@@ -4,12 +4,18 @@
     "strict": true,
     "lib": ["deno.ns", "deno.unstable", "deno.worker"]
   },
+  "tasks": {
+    "comment1": "echo cwd is by default the directory of deno.json",
+    "comment2": "echo cannot restrict ffi to a lib https://github.com/denoland/deno/issues/15511",
+    "run": "cd .. && deno run --config=typegate/deno.json --unstable --allow-run=hostname --allow-sys --allow-env --allow-hrtime --allow-write=tmp --allow-ffi --allow-read=. --allow-net typegate/src/main.ts",
+    "test": "cd .. && deno test --config=typegate/deno.json --unstable --allow-run=hostname,target/debug/meta --allow-sys --allow-env --allow-hrtime --allow-write=tmp,typegate/tests --allow-ffi --allow-read=. --allow-net"
+  },
   "lock": "deno.lock",
   "imports": {
     "std/": "https://deno.land/std@0.170.0/",
     "graphql": "https://cdn.skypack.dev/graphql@16.6.0?dts",
     "graphql/ast": "https://cdn.skypack.dev/graphql@16.6.0/language/ast?dts",
-    "sentry": "npm:@sentry/node@7.12.1",
+    "sentry": "npm:@sentry/node@7.38.0",
     "dataloader": "npm:dataloader@2.1.0",
     "chance": "npm:chance@1.1.8",
     "swc": "https://deno.land/x/swc@0.2.1/mod.ts",
