+++ typegate/native/Cargo.toml
@@ -38,7 +38,7 @@ serde_json = "1.0.93"
 dashmap = "5.4.0"
 common = { path = "../../libs/common" }
 macros = { path = "../../libs/macros" }
-sentry = "0.29.3"
+sentry = { version = "0.29.3", features = ["anyhow"] }
 tempfile = "3.3.0"
 convert_case = "0.6.0"
 rust-s3 = { git = "https://github.com/zifeo/rust-s3", branch = "order" }
