+++ typegate/native/src/lib.rs
@@ -1,6 +1,7 @@
 // Copyright Metatype OÜ under the Elastic License 2.0 (ELv2). See LICENSE.md for usage.
 
 mod config;
+mod errors;
 mod runtimes;
 mod typegraph;
