+++ typegate/src/typegraph.ts
@@ -52,7 +52,7 @@ export { Cors, Rate, TypeGraphDS, TypeMaterializer, TypePolicy, TypeRuntime };
 
 export type RuntimeResolver = Record<string, Runtime>;
 
-const logger = getLogger(import.meta);
+const _logger = getLogger(import.meta);
 
 const runtimeInit: RuntimeInit = {
   s3: S3Runtime.init,
@@ -183,7 +183,7 @@ export class TypeGraph {
           }`,
         );
 
-        logger.debug(`init ${runtime.name} (${idx})`);
+        //logger.debug(`init ${runtime.name} (${idx})`);
         return runtimeInit[runtime.name]({
           typegraph,
           materializers: typegraph.materializers.filter(
@@ -209,11 +209,11 @@ export class TypeGraph {
 
   async deinit(): Promise<void> {
     for await (
-      const [idx, runtime] of this.runtimeReferences.map(
+      const [_idx, runtime] of this.runtimeReferences.map(
         (rt, i) => [i, rt] as const,
       )
     ) {
-      logger.debug(`deinit runtime ${idx}`);
+      //logger.debug(`deinit runtime ${idx}`);
       await runtime.deinit();
     }
     if (this.introspection) {
