+++ website/.env.sample
@@ -1,12 +1,12 @@
-TG_DOCS_SENDINBLUE_KEY=""
-TG_HOMEPAGE_GITHUB_CLIENT_ID=""
-TG_HOMEPAGE_GITHUB_CLIENT_SECRET=""
+TG_DOCS_SENDINBLUE_KEY="fake"
+TG_HOMEPAGE_GITHUB_CLIENT_ID="fake"
+TG_HOMEPAGE_GITHUB_CLIENT_SECRET="fake"
 TG_HOMEPAGE_POSTGRES_CONN="postgresql://postgres:password@localhost:5432/db?schema=homepage"
 TG_ORM_FOR_THE_EDGE_POSTGRES_CONN="postgresql://postgres:password@localhost:5432/db?schema=homepage"
 TG_PRISMA_POSTGRES_CONN="postgresql://postgres:password@localhost:5432/db?schema=homepage"
 TG_DATABASE_POSTGRES_CONN="postgresql://postgres:password@localhost:5432/db?schema=homepage"
 TG_POLICIES_BASIC_ADMIN="admin_pass"
 TG_POLICIES_BASIC_USER="user_pass"
-TG_AUTH_GITHUB_CLIENT_ID=""
-TG_AUTH_GITHUB_CLIENT_SECRET=""
+TG_AUTH_GITHUB_CLIENT_ID="fake"
+TG_AUTH_GITHUB_CLIENT_SECRET="fake"
 TG_AUTH_BASIC_ADMIN="pass"
