+++ website/docs/reference/changelog.mdx
@@ -338,6 +338,27 @@ comments: false
 
 <!-- vale off -->
 
+## [v0.0.3-dev.5](https://github.com/metatypedev/metatype/releases/tag/v0.0.3-dev.5) (2/24/2023)
+
+<!-- Release notes generated using configuration in .github/release.yml at v0.0.3-dev.5 -->
+
+### What's Changed
+
+#### New features 🎉
+
+* chore(deps): bump temporal-sdk-core-protos from `af26dee` to `0d7908c` by @dependabot in [#142](https://github.com/metatypedev/metatype/pull/142)
+* feat: add type node either by @msga-mmm in [#110](https://github.com/metatypedev/metatype/pull/110)
+* chore: prepare release 0.0.3-dev.5 by @github-actions in [#146](https://github.com/metatypedev/metatype/pull/146)
+* chore(deps): bump temporal-client from `0d7908c` to `09a2e4f` by @dependabot in [#156](https://github.com/metatypedev/metatype/pull/156)
+* chore(deps): bump syn from 1.0.107 to 1.0.108 by @dependabot in [#155](https://github.com/metatypedev/metatype/pull/155)
+* chore(deps): bump swc_ecma_codegen from 0.129.8 to 0.129.10 by @dependabot in [#153](https://github.com/metatypedev/metatype/pull/153)
+* chore(deps): bump swc_ecma_transforms_typescript from 0.164.10 to 0.164.11 by @dependabot in [#154](https://github.com/metatypedev/metatype/pull/154)
+* chore(deps): bump python-box from 7.0.0 to 7.0.1 in /typegraph by @dependabot in [#151](https://github.com/metatypedev/metatype/pull/151)
+* feat: better logging and fix deployment by @zifeo in [#150](https://github.com/metatypedev/metatype/pull/150)
+
+
+**Full Changelog**: [v0.0.3-dev.4...v0.0.3-dev.5](https://github.com/metatypedev/metatype/compare/v0.0.3-dev.4...v0.0.3-dev.5)
+
 ## [v0.0.3-dev.4](https://github.com/metatypedev/metatype/releases/tag/v0.0.3-dev.4) (2/22/2023)
 
 <!-- Release notes generated using configuration in .github/release.yml at v0.0.3-dev.4 -->
